requirements.txt
Django==1.11.14
django-classy-tags==0.8.0
django-cms==3.5.2
django-formtools==2.1
django-sekizai==0.10.0
django-treebeard==4.3
djangocms-admin-style==1.2.8
djangocms-text-ckeditor==3.6.0
html5lib==0.9999999
Pillow==5.1.0
psycopg2==2.7.4
pytz==2018.5
six==1.11.0
radar==0.3